Refactor GUI into modular package
