+++ talkmatch/gui/__init__.py
@@ -1,0 +1,6 @@
+"""GUI components for TalkMatch."""
+
+from .chat_box import ChatBox
+from .control_panel import ControlPanel, run_app
+
+__all__ = ["ChatBox", "ControlPanel", "run_app"]
